Move CustomHoverGestureRecognizer to separate file
